Changes to Graph class
From HashTable to HashMap. HashTable is thread safe, but slower. We only have a single thread so no need for it, and HashMap has a better implementation when handling collision
addEdge added a check to make sure we only set an edge for vertex we actually added to our Graph.
showConnection better for looping on the map, no need to create an additional array and unnecessary conversion.